Added PlaceOrder and RemoveOrder Sagas.

+++ OrderMicroservice/OrderService.Core/Interfaces/IOrderRepository.cs
@@ -10,7 +10,7 @@ public interface IOrderRepository
     public Task<Order> CreateAsync(Order order);
     public Task<Order> UpdateAsync(Guid id, OrderStatus status);
     public Task<Order> DeleteAsync(Guid id);
-    public Task AddItemToOrder(Guid orderId, int productId, string productName, double price, int quantity);
+    public Task AddItemToOrder(Guid orderId, int productId, string productName, decimal price, int quantity);
     public Task DeleteOrderItemFromOrder(Guid orderId, int productId, int quantity);
     public Task<Order?> GetActiveOrderByUserId(int i);
 }
