+++ PaymentMicroservice/PaymentService.Application/Services/PaymentService.cs
@@ -13,7 +13,7 @@ private readonly IPaymentRepository _paymentRepository;
         _paymentRepository = paymentRepository;
     }
 
-    public bool CreatePayment(Guid orderId, int userId, double amount)
+    public bool CreatePayment(Guid orderId, int userId, decimal amount)
     {
         var random = new Random();
         if (random.Next(1, 10) < 8)
